chore: organaze directory
