+++ lib/model/weather.dart
@@ -1,18 +1,3 @@
-import 'package:yumemi_weather/yumemi_weather.dart';
-
-class WeatherResult {
-  /// コンストラクタ
-  WeatherResult({
-    required this.successed,
-    this.weather,
-    this.error,
-  });
-
-  bool successed;
-  Weather? weather;
-  YumemiWeatherError? error;
-}
-
 class Weather {
   Weather({
     required this.weatherCondition,
