+++ lib/model/weather_result.dart
@@ -1,0 +1,15 @@
+import 'package:flutter_training/model/weather.dart';
+import 'package:yumemi_weather/yumemi_weather.dart';
+
+class WeatherResult {
+  /// コンストラクタ
+  WeatherResult({
+    required this.successed,
+    this.weather,
+    this.error,
+  });
+
+  bool successed;
+  Weather? weather;
+  YumemiWeatherError? error;
+}
